ENG-10259: remove client, update unit tests
