+++ voltdbclient/authentication_test.go
@@ -23,13 +23,11 @@ import (
 )
 
 func TestLoginRequest(t *testing.T) {
-	config := ClientConfig{"hello", "world"}
 	var loginBytes []byte
 	loginBuf := bytes.NewBuffer(loginBytes)
-	client := Client{&config, nil, loginBuf, nil, nil, 0}
-	login, err := serializeLoginMessage(client.config.username, client.config.password)
+	login, err := serializeLoginMessage("hello", "world")
 	check(t, err)
-	writeLoginMessage(client.getWriter(), &login)
+	writeLoginMessage(loginBuf, &login)
 
 	fileBytes, err := ioutil.ReadFile("./test_resources/authentication_request_sha256.msg")
 	check(t, err)
@@ -43,9 +41,7 @@ func TestLoginResponse(t *testing.T) {
 	b, err := ioutil.ReadFile("./test_resources/authentication_response.msg")
 	check(t, err)
 	reader := bytes.NewReader(b)
-	config := ClientConfig{"", ""}
-	client := Client{&config, reader, nil, nil, nil, 0}
-	connData, err := readLoginResponse(client.getReader())
+	connData, err := readLoginResponse(reader)
 	check(t, err)
 	// consider as passed if returns non nil connection data
 	if connData == nil {
