+++ voltdbclient/client_write_test.go
@@ -18,30 +18,31 @@ package voltdbclient
 
 import (
 	"bytes"
+	"database/sql/driver"
 	"fmt"
-	"io/ioutil"
+	//"io/ioutil"
 	"testing"
+	//"time"
+	"io/ioutil"
 	"time"
 )
 
 func TestSimpleProcedureCall(t *testing.T) {
 
 	// this is the writer, write the serialized procedure to this buffer.
-	config := ClientConfig{"", ""}
 	var bs []byte
 	buf := bytes.NewBuffer(bs)
-	client := Client{&config, nil, buf, nil, nil, 0}
+	conn := VoltConn{nil, buf, nil, nil, nil, nil, true}
 	var handle int64 = 51515
-	client.writeProcedureCall(client.writer, "HELLOWORLD.insert", handle, []interface{}{"Bonjour", "Monde", "French"})
+	conn.serializeQuery(buf, "HELLOWORLD.insert", handle, []driver.Value{"Bonjour", "Monde", "French"})
 	r := bytes.NewReader(buf.Bytes())
 	checkSimpleBuffer(t, r, 0, "HELLOWORLD.insert", 51515, 3, "Bonjour", "Monde", "French")
 }
 
 func TestInsertDifferentTypes(t *testing.T) {
-	config := ClientConfig{"", ""}
 	var bs []byte
 	buf := bytes.NewBuffer(bs)
-	client := Client{&config, nil, buf, nil, nil, 0}
+	conn := VoltConn{nil, buf, nil, nil, nil, nil, true}
 
 	var id int32 = 100
 	var nid int32 = 100
@@ -59,7 +60,7 @@ func TestInsertDifferentTypes(t *testing.T) {
 	now := time.Date(2016, time.June, 10, 23, 0, 0, 0, time.UTC)
 
 	var handle int64 = 61616
-	client.writeProcedureCall(client.writer, "EXAMPLE_OF_TYPES.insert", handle, []interface{}{id, nid, name, data, status, typ, pan, bo, now})
+	conn.serializeQuery(conn.writer, "EXAMPLE_OF_TYPES.insert", handle, []driver.Value{id, nid, name, data, status, typ, pan, bo, now})
 
 	// read the verification file into a buffer and compare the two buffers
 	bs, err := ioutil.ReadFile("./test_resources/verify_insert_types.msg")
